Added new admin role, route to read and delete users data

+++ fastapienv/TodoApp/routers/todos.py
@@ -50,7 +50,7 @@ async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Pat
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail="Todo not found")
-# Tämä endpoint palauttaa yksittäisen todo-tietueen tietokannasta.
+# Tämä endpoint palauttaa yksittäisen todo-tietueen tietokannasta jos olet sisäänkirjautunut / oikeutettu.
 
 @router.post("/todo", status_code=201)
 async def create_todo(user: user_dependency, db: db_dependency, 
@@ -68,10 +68,14 @@ async def create_todo(user: user_dependency, db: db_dependency,
 
 
 @router.put("/todo/{todo_id}", status_code=204)
-async def update_todo(db: db_dependency,
-                      todo_id:int,
-                      todo_request: TodoRequest):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user: user_dependency, db: db_dependency,
+                      todo_request: TodoRequest,
+                      todo_id:int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+    #tarkistetaan, että käyttäjä on kirjautunut sisään
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found.')
     
@@ -85,11 +89,16 @@ async def update_todo(db: db_dependency,
 # Tämä endpoint päivittää olemassa olevan todo-tietueen tietokannassa.
 
 @router.delete("/todo/{todo_id}", status_code=204)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+        
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
+    
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found.')
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
 
     db.commit()
 # Tämä endpoint poistaa todo-tietueen tietokannasta.
